archive/newproject/requirements.txt.txt
PyYAML = 6.0
pandas = 1.4.2
scikit-learn = 1.0.2
tensorflow = 2.9.1
keras = 2.9.0
pillow = 9.0.1
opencv-python = 4.8.0.74